+++ src/components/Header/components/navigation/index.tsx
@@ -7,7 +7,7 @@ import { Store } from '@/store'
 import './navigation.scss'
 
 export default function Navigation() {
-    const userStore = useSelector((store: Store) => store.userStore)
+    const candidateStore = useSelector((store: Store) => store.candidateStore)
     const navigate = useNavigate()
 
     return (
@@ -39,7 +39,7 @@ export default function Navigation() {
 
                 {/* BUTTON GROUP */}
                 {
-                    userStore.data ? (
+                    candidateStore.data ? (
                         <>
                             {/* bell */}
                             <svg width="19" height="21" viewBox="0 0 19 21" fill="none" xmlns="http://www.w3.org/2000/svg">
@@ -58,7 +58,7 @@ export default function Navigation() {
                             </svg>
 
                             {/* name user */}
-                            <div className='user_name'>{userStore.data.username}</div>
+                            <div className='user_name'>{candidateStore.data.name}</div>
                         </>
                     ) : (
                         <div className='button-group'>
@@ -85,7 +85,7 @@ export default function Navigation() {
 
 
             {
-                userStore.data ? (
+                candidateStore.data ? (
                     <>
                         <Navigate />
                     </>
