+++ src/components/authen/login/candidate/candidate_login.scss
@@ -1,0 +1,138 @@
+* {
+    font-family: cabin;
+}
+
+.box-logo-rikkei {
+    padding: 60px;
+    width: 100%;
+
+    img {
+        Width: 200px;
+        Height: 67px;
+        cursor: pointer;
+    }
+}
+
+.box-content {
+    max-width: 100%;
+    width: 100%;
+    justify-content: center;
+    align-items: center;
+    display: flex;
+
+    .box-content-left {
+        width: 40%;
+
+        form {
+            width: 100%;
+            margin-bottom: 157px;
+            border-right: 2px solid;
+            border-image-slice: 1;
+            border-image-source: linear-gradient(to bottom, #ffffff, #0A033C, #ffffff);
+            padding: 64px;
+
+            .title-rikkei {
+                width: 100%;
+
+                h3 {
+                    width: 470px;
+                    height: 106px;
+                    font-size: 30px;
+                }
+            }
+
+            .box-item-content {
+                margin-top: 50px;
+
+                label {
+                    font-size: 16px;
+                }
+
+                .input-email {
+                    width: 486px;
+                    height: 65px;
+                    width: 100%;
+                    // padding: 10px 10px;
+                    margin-top: 15px;
+
+                    &::placeholder {
+                        font-size: 14px;
+                        font-weight: 400;
+                        color: #5D5A6F;
+                    }
+                }
+
+                .input-password {
+                    width: 486px;
+                    height: 65px;
+                    // padding: 10px 10px;
+                    cursor: pointer;
+                    width: 100%;
+                    margin-bottom: 20px;
+                    margin-top: 15px;
+
+                    &::placeholder {
+                        font-size: 14px;
+                        font-weight: 400;
+                        color: #5D5A6F;
+                    }
+                }
+            }
+
+            .box-item-forgot {
+                font-size: 14px;
+                color: #5D5A6F;
+                cursor: pointer;
+                font-weight: 500;
+                width: 100%;
+                margin-top: 20px;
+                margin-bottom: 20px;
+                text-align: center;
+            }
+
+            .box-item-create-account {
+                font-weight: 500;
+                font-size: 14px;
+                text-align: center;
+                color: #5D5A6F;
+
+                span {
+                    cursor: pointer;
+                    color: #BC2228;
+                }
+            }
+
+
+            .box-button-login {
+                .button-login {
+                    cursor: pointer;
+                    font-size: 16px;
+                    width: 486px;
+                    height: 65px;
+                    width: 100%;
+                    color: #ffffff;
+                    border-radius: 5px;
+                    border: none;
+                    outline: none;
+                    padding: 10px 10px;
+                    background-color: #BC2228;
+                }
+            }
+
+        }
+    }
+
+    .box-content-right {
+        .box-img-content {
+            width: 100%;
+            padding: 10px;
+
+            img {
+                width: 500px;
+                height: 500px;
+            }
+        }
+    }
+
+
+}
