+++ src/main.tsx
@@ -1,6 +1,10 @@
 import ReactDOM from 'react-dom/client'
 import App from './App.tsx'
+import { store } from '@/store/index.ts'
+import { Provider } from 'react-redux'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
-    <App />
+    <Provider store={store} >
+        <App />
+    </Provider>
 )
