+++ src/utils/lazies/components/Loading.tsx
@@ -1,0 +1,15 @@
+
+import Icon from "./loading.png"
+
+import "./loading.scss"
+
+export default function Loading() {
+  return (
+    <div className="loading_container">
+      <img
+        className="rotating-image"
+        src={Icon}
+      />
+    </div>
+  )
+}
