+++ src/utils/lazies/components/loading.scss
@@ -1,0 +1,28 @@
+.loading_container {
+    width: 100%;
+    height: 100vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: rgba(white, 0.7);
+    position: fixed;
+    top: 0;
+    left: 0;
+    z-index: 10000;
+    .rotating-image {
+      width: 250px;
+      height: 250px;
+      position: relative;
+      animation: rotate 2s infinite linear;
+    }
+  }
+  
+  @keyframes rotate {
+    from {
+      transform: rotate(0deg);
+    }
+    to {
+      transform: rotate(360deg);
+    }
+  }
+
